requirements.txt
asgiref==3.6.0
Django==3.2
django-allauth==0.41.0
django-crispy-forms==1.12.0
oauthlib==3.2.2
Pillow==9.4.0
python3-openid==3.2.0
pytz==2022.7.1
requests-oauthlib==1.3.1
sqlparse==0.4.3
stripe==5.1.1
